pyproject.toml
[tool.poetry]
name = "goit-pythonweb-hw-12"
version = "0.1.0"
description = ""
authors = ["i3oi3ka <i3oi3ka@ukr.net>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.13"
sqlalchemy = "^2.0.43"
fastapi = "^0.116.1"
alembic = "^1.16.5"
python-dotenv = "^1.1.1"
email-validator = "^2.3.0"
asyncpg = "^0.30.0"
uvicorn = "^0.35.0"
pydantic-settings = "^2.10.1"
slowapi = "^0.1.9"
passlib = "^1.7.4"
fastapi-mail = "^1.5.0"
python-jose = "^3.5.0"
python-multipart = "^0.0.20"
bcrypt = "<4.0.0"
cloudinary = "^1.44.1"
libgravatar = "^1.0.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
